sysconf:git-subtree

+++ install.sh
@@ -2,18 +2,29 @@
 
 . /usr/lib/sysconf.base/common.sh
 
+# Debian wheezy does not have git-subtree
+if [ ! -f /usr/lib/git-core/subtree ]; then
+    cp resource/git-subtree /usr/lib/git-core/subtree \
+        && chmod +x /usr/lib/git-core/subtree \
+        || nef_fatal "could not install git-subtree"
+fi
+
+# Install NodeJS and fix bon/node -> nodejs
 sysconf_require_packages nodejs \
     || nef_fatal "could not install nodejs"
-
 ln -s nodejs /usr/bin/node
 
-sh install.npm.sh \
-    || nef_fatal "could not install npm"
+# Install NPM
+if [ ! -f /usr/bin/npm]; then
+    curl https://www.npmjs.org/install.sh | sh \
+        || nef_fatal "could not install npm"
+fi
 
-# sed -i 's/env node$/env nodejs/' /usr/bin/npm
-
+# 'forever' come from NPM and provide NodeJS process management
 npm install -g forever \
     || nef_fatal "could not install npm module: forever"
+
+# Manage the glxc-http service
 
 update-rc.d glxc-http defaults
